+++ CHANGELOG.md
@@ -10,6 +10,10 @@
 - Fixes: added v0.3 draft documentation link & updates shields in `readme.md`
 - Fixes: generated web components will always end with `-component`
 
+### Removed
+
+- Removes: `cross-env` dependency
+
 ## 0.3.0-a1 - 2019-08-19
 
 ### Added